requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
av==9.2.0
backcall==0.2.0
bcrypt==4.0.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
configobj==5.0.6
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
cytoolz==0.12.0
debugpy==1.6.3
decorator==5.1.1
decord==0.6.0
easydict==1.10
einops==0.4.1
entrypoints==0.4
executing==1.1.0
fairscale==0.4.12
fastapi==0.85.0
ffmpy==0.3.0
filelock==3.8.0
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
google-auth==2.12.0
google-auth-oauthlib==0.4.6
gradio==3.4.0
grpcio==1.49.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.9.1
idna==3.4
imageio==2.22.1
importlib-metadata==5.0.0
inflect==6.0.0
ipdb==0.13.9
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
linkify-it-py==1.0.3
lmdb==1.3.0
lz4==4.0.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
msgpack==1.0.4
msgpack-numpy==0.4.8
multidict==6.0.2
nest-asyncio==1.5.6
networkx==2.8.7
nltk==3.7
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
orjson==3.8.0
packaging==21.3
pandas==1.5.0
paramiko==2.11.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pillow-SIMD==9.0.0.post1
prettytable==3.4.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydub==0.25.1
Pygments==2.13.0
pymongo==4.2.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.base==0.3.0
ruamel.yaml.clib==0.2.6
ruamel.yaml.cmd==0.6.3
ruamel.yaml.convert==0.3.2
sacremoses==0.0.53
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
sniffio==1.3.0
stack-data==0.5.1
starlette==0.20.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tifffile==2022.8.12
timm==0.6.7
tokenizers==0.10.3
toml==0.10.2
toolz==0.12.0
torch==1.10.0+cu113
torchvision==0.11.0+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.11.3
typing_extensions==4.3.0
uc-micro-py==1.0.1
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
yacs==0.1.8
yarl==1.8.1
zipp==3.9.0
